refactor(T5 U2): impl-split SmartUpgradePanel across 2 .cpp [BUILD-PASSED]
Slice U2. Pure impl-split (one UWidget class, 2 .cpp); .h BYTE-IDENTICAL so the Smart_UpgradePanel_Widget
BP contract is preserved. _Detail.cpp (1,172: row selection + cost display/calc + tabs + triage +
traversal). SmartUpgradePanel.cpp 2,138 -> 1,010. Each TU has its own #define/#undef LOCTEXT_NAMESPACE.
Shipping PackagePlugin clean (ExitCode=0). Solo lane; recompile widget BP at next deploy (no-op).
